MYR-71 Split audit.go and hub.go to satisfy 300-line file cap
CLAUDE.md "File Rules" caps non-test files at 300 lines. PR #216
review (Claude AI) flagged internal/mask/audit.go (393) and
internal/ws/hub.go (412) as violations. Split with no behavior
change:

- internal/mask/audit.go (246 lines): types, AuditEmitter,
  AuditMetrics, NoopAuditMetrics, BuildEntry, newAuditID,
  allowedMetadataKeysIntersect.
- internal/mask/audit_emit.go (99 lines): EmitAsync +
  emitDetached + auditInsertTimeout.
- internal/mask/audit_sampling.go (60 lines): ShouldAuditREST,
  ShouldAuditWS, writeField, auditSampleModulus.

- internal/ws/hub.go (164 lines): keeps Hub struct, NewHub,
  Register/Unregister/Broadcast/BroadcastAll/Stop/ClientCount/
  ipConnectionCount.
- internal/ws/hub_masked.go (259 lines): WithMaskAudit,
  nextFrameSeq, BroadcastMasked, buildRoleFrames,
  maybeEmitAuditWS, marshalVehicleUpdate, v1Roles. The split
  lines up exactly with the MYR-71 surface area, so git blame
  on the existing fields stays clean.

Also adds a short comment on hub_masked.go's nextFrameSeq
documenting that the unbounded sync.Map is acceptable in v1
(small vehicle set; DV-02 retires the counter entirely).
