+++ internal/mask/audit.go
@@ -3,31 +3,14 @@ package mask
 import (
 	"context"
 	"crypto/rand"
-	"encoding/binary"
 	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"hash/fnv"
-	"io"
-	"log/slog"
-	"strconv"
 	"time"
 
 	"github.com/tnando/my-robo-taxi-telemetry/internal/auth"
 )
-
-// Sampling rate for the mask audit log. Per docs/contracts/rest-api.md
-// §5.3, every masked response (REST or WS) MUST be audit-logged at a
-// 1% deterministic rate computed by hash modulo 100.
-const auditSampleModulus uint64 = 100
-
-// auditInsertTimeout caps the time a fire-and-forget Emit goroutine
-// will wait on an InsertAuditLog call. The hot path (BroadcastMasked
-// fan-out, REST response write) MUST NOT block on this; the goroutine
-// uses a detached context with this timeout so a stuck DB pool cannot
-// pile up audit goroutines forever.
-const auditInsertTimeout = 2 * time.Second
 
 // AuditChannel labels the audit row's transport. The set is closed at
 // the two transports a v1 mask projection can flow through.
@@ -224,91 +207,6 @@ func BuildEntry(
 	}, nil
 }
 
-// EmitAsync is the fire-and-forget wrapper around AuditEmitter.
-// Invariants required by the mask audit pipeline (rest-api.md §5.3,
-// data-lifecycle.md §4):
-//
-//   - Hot-path non-blocking: failures MUST NOT drop the masked frame
-//     or the REST response. EmitAsync returns immediately to the
-//     caller; the actual insert runs on a spawned goroutine.
-//   - Detached context: the caller's request context may be canceled
-//     mid-write (e.g., the HTTP client closes the connection). We
-//     use context.WithoutCancel + a 2 s timeout so the audit row
-//     still lands even after the caller's context dies.
-//   - Bounded latency: the 2 s timeout caps how long a stuck DB pool
-//     can keep an audit goroutine alive. Beyond that we increment the
-//     failure metric, log slog.Warn, and drop the row.
-//   - Metric coverage: every successful insert increments
-//     audit_log_writes_total{action, target}; every failure increments
-//     audit_log_write_failures_total{action, target}. The labels match
-//     the tuple of allow-listed enum values, so cardinality stays
-//     bounded.
-//
-// EmitAsync is a no-op if emitter is nil — the production wiring may
-// pass nil before the AuditRepo is composed (e.g., during dev mode
-// when the writer is intentionally disabled), and a no-op keeps the
-// hot path quiet rather than logging a flood of "audit emitter not
-// configured" warnings.
-func EmitAsync(
-	parent context.Context,
-	emitter AuditEmitter,
-	metrics AuditMetrics,
-	logger *slog.Logger,
-	entry AuditEntry,
-) {
-	if emitter == nil {
-		return
-	}
-	if metrics == nil {
-		metrics = NoopAuditMetrics{}
-	}
-	go emitDetached(parent, emitter, metrics, logger, entry)
-}
-
-// emitDetached runs the actual insert with a detached context. Split
-// out so EmitAsync remains trivial and tests can drive it directly.
-func emitDetached(
-	parent context.Context,
-	emitter AuditEmitter,
-	metrics AuditMetrics,
-	logger *slog.Logger,
-	entry AuditEntry,
-) {
-	defer func() {
-		// Recover from unexpected panics inside the emitter so a buggy
-		// implementation cannot crash the hot path's goroutine.
-		if r := recover(); r != nil {
-			metrics.IncAuditWriteFailure(entry.Action, entry.TargetType)
-			if logger != nil {
-				logger.Warn("mask.EmitAsync: emitter panicked",
-					slog.Any("recover", r),
-					slog.String("action", entry.Action),
-					slog.String("target_type", entry.TargetType),
-				)
-			}
-		}
-	}()
-
-	// Detach from the caller's request context — once we decide to
-	// emit, we follow through even if the caller cancels.
-	ctx := context.WithoutCancel(parent)
-	ctx, cancel := context.WithTimeout(ctx, auditInsertTimeout)
-	defer cancel()
-
-	if err := emitter.InsertAuditLog(ctx, entry); err != nil {
-		metrics.IncAuditWriteFailure(entry.Action, entry.TargetType)
-		if logger != nil {
-			logger.Warn("mask.EmitAsync: insert failed",
-				slog.String("action", entry.Action),
-				slog.String("target_type", entry.TargetType),
-				slog.String("error", err.Error()),
-			)
-		}
-		return
-	}
-	metrics.IncAuditWrite(entry.Action, entry.TargetType)
-}
-
 // auditIDPrefix is the leading character on every mask audit cuid
 // (matches the convention used elsewhere in the project's Prisma
 // tables — every cuid starts with `c`).
@@ -345,49 +243,4 @@ func allowedMetadataKeysIntersect(keys []string) bool {
 		}
 	}
 	return true
-}
-
-// ShouldAuditREST returns true when this REST response should be
-// emitted to the audit log. The decision is deterministic given the
-// inputs — replaying the same userID + requestID + resourceID will
-// always return the same boolean. Per rest-api.md §5.3, the inputs are
-// joined with a separator before hashing to avoid collisions across
-// distinct triples that share a concatenated form.
-func ShouldAuditREST(userID, requestID, resourceID string) bool {
-	h := fnv.New64a()
-	writeField(h, userID)
-	writeField(h, requestID)
-	writeField(h, resourceID)
-	return h.Sum64()%auditSampleModulus == 0
-}
-
-// ShouldAuditWS returns true when this WebSocket frame should be
-// audit-logged. Per rest-api.md §5.3, the WS audit emit is per
-// (vehicleID, role, frame) at the hub layer (NOT per client) — the
-// hash inputs reflect that scope.
-//
-// frameSeq SHOULD be the envelope sequence number once DV-02 lands.
-// Until then, callers can pass a per-vehicle monotonic counter — the
-// determinism only requires that the counter is reproducible during
-// replay.
-func ShouldAuditWS(vehicleID string, role auth.Role, frameSeq uint64) bool {
-	h := fnv.New64a()
-	writeField(h, vehicleID)
-	writeField(h, string(role))
-	var buf [8]byte
-	binary.BigEndian.PutUint64(buf[:], frameSeq)
-	_, _ = h.Write(buf[:])
-	return h.Sum64()%auditSampleModulus == 0
-}
-
-// writeField writes a length-prefixed string into the hash. The length
-// prefix prevents ambiguity between (e.g.) {"abc", "def"} and
-// {"ab", "cdef"} which would otherwise hash to the same byte sequence
-// when concatenated. `hash.Hash64` already implements `io.Writer`, so
-// the hasher is passed in directly without an inline anonymous
-// interface (PR #195 review suggestion #2).
-func writeField(h io.Writer, s string) {
-	_, _ = h.Write([]byte(strconv.Itoa(len(s))))
-	_, _ = h.Write([]byte(":"))
-	_, _ = h.Write([]byte(s))
 }
